src/requirements.txt
alembic==1.4.2
bs4
click==7.1.1
discord.py
Flask==1.1.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
python-dateutil==2.8.1
python-editor==1.0.4
pyyaml
requests
six==1.14.0
SQLAlchemy==1.3.15
Werkzeug==1.0.0
WTForms==2.2.1
